+++ src/io/parser/obj.rs
@@ -69,6 +69,7 @@ impl Loaded {
                         1.0
                     },
                     normal_texture,
+                    lighting_model: LightingModel::Blinn,
                     ..Default::default()
                 });
             }
