+++ src/renderer.rs
@@ -4,7 +4,8 @@
 //!
 
 pub use crate::core::{
-    math::*, render_states::*, render_target::*, texture::*, Camera, Context, Viewport,
+    math::*, render_states::*, render_target::*, texture::*, Camera, Context, LightingModel,
+    Viewport,
 };
 
 pub mod material;
