+++ src/renderer/light/environment.rs
@@ -1,5 +1,4 @@
 use crate::core::*;
-use crate::renderer::*;
 
 ///
 /// Precalculations of light shining from an environment map (known as image based lighting - IBL).
